added hittable fault constructor

+++ SummerNodeProject/HashTable.h
@@ -16,10 +16,11 @@ template<class Type>
 class HashTable
 {
 private:
-    int capacity;  // can hold
-    int size;      // items in the table
+    long capacity;  // can hold
+    long size;      // items in the table
     double efficiencyPercentage;  // how full do you want it to get before resizing
-    HashNode<Type> * front;
+    HashNode<Type> * front;    // points to entire array
+    
     bool isPrime(int number);
     void resize();
     long nextPrime(long current);
